use ddl hook instead of go failpoint
